moved to named_entities.py
