--- package/kedro_viz/models/flowchart/pipeline.py
@@ -1,7 +1,0 @@
-"""`kedro_viz.models.flowchart.pipelines` represent Kedro pipelines in a viz graph."""
-
-from .model_utils import NamedEntity
-
-
-class RegisteredPipeline(NamedEntity):
-    """Represent a registered pipeline in a Kedro project."""
