+++ package/tests/test_data_access/test_managers.py
@@ -16,13 +16,13 @@ from kedro_viz.data_access.repositories.modular_pipelines import (
 )
 from kedro_viz.integrations.utils import UnavailableDataset
 from kedro_viz.models.flowchart.edge import GraphEdge
+from kedro_viz.models.flowchart.named_entities import Tag
 from kedro_viz.models.flowchart.nodes import (
     DataNode,
     ParametersNode,
     TaskNode,
     TranscodedDataNode,
 )
-from kedro_viz.models.flowchart.tag import Tag
 
 
 def identity(x):
